FINALLY GOT DATA DUMPS WORKING
So, the problem was in the fact that yaml wasn't storing PartMachine's
list of parts, because the lsit of parts is NEVER INITIALIZED IN IT.
Yaml will only store the object variables that are initialized in the
init function. That is absurdly stupid. So I can use a variable for
years by OH WAIT DIDN'T PUT IT IN INIT I GUESS IT DOESN'T EXIST I'M YAML
AND I'M DUMB DURRRRRRRRRRRR.....